Library is building finally. Starting implementation.
